Fragments
Create fragment file for the four categories

Transfer most of the code from the activity files to the fragment files
