+++ app/src/main/java/com/example/android/miwok/ColorsActivity.java
@@ -32,75 +32,82 @@ import java.util.ArrayList;
 
 public class ColorsActivity extends AppCompatActivity {
 
-    private MediaPlayer mMediaPlayer;
+//    private MediaPlayer mMediaPlayer;
+//
+//    private MediaPlayer.OnCompletionListener mCompletionListener =  new MediaPlayer.OnCompletionListener(){
+//        @Override
+//        public void onCompletion(MediaPlayer mediaPlayer) {
+//            releaseMediaPlayer();
+//        }
+//    };
 
-    private MediaPlayer.OnCompletionListener mCompletionListener =  new MediaPlayer.OnCompletionListener(){
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            releaseMediaPlayer();
-        }
-    };
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.word_list);
 
-        final ArrayList<Word> words = new ArrayList<Word>();
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.container, new ColorsFragment())
+                .commit();
 
-//        words.add("one");
-        words.add(new Word("red", "wetetti", R.drawable.color_red, R.raw.color_red));
-        words.add(new Word("green", "chokokki", R.drawable.color_green, R.raw.color_green));
-        words.add(new Word("brown", "takaakki", R.drawable.color_brown, R.raw.color_brown));
-        words.add(new Word("gray", "topoppi", R.drawable.color_gray, R.raw.color_gray));
-        words.add(new Word("black", "kululli", R.drawable.color_black, R.raw.color_black));
-        words.add(new Word("white", "kelelli", R.drawable.color_white, R.raw.color_white));
-        words.add(new Word("dusty yellow", "topiise", R.drawable.color_dusty_yellow, R.raw.color_dusty_yellow));
-        words.add(new Word("mustard yellow", "chiwiite", R.drawable.color_mustard_yellow, R.raw.color_mustard_yellow));
-
-        WordAdapter adapter = new WordAdapter(this, words, R.color.category_colors);
-
-        ListView listView = (ListView) findViewById(R.id.list);
-
-        listView.setAdapter(adapter);
-
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
-                Word word = words.get(position);
-
-                releaseMediaPlayer();
-
-                mMediaPlayer = MediaPlayer.create(ColorsActivity.this, word.getmAudioResourceId());
-                mMediaPlayer.start(); // no need to call prepare(); create() does that for you
-
-                mMediaPlayer.setOnCompletionListener(mCompletionListener);
-            }
-        });
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        // When the activity is stopped, release the media player resources because we won't
-        // be playing any more sounds.
-        releaseMediaPlayer();
-    }
-
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-        }
+//        final ArrayList<Word> words = new ArrayList<Word>();
+//
+////        words.add("one");
+//        words.add(new Word("red", "wetetti", R.drawable.color_red, R.raw.color_red));
+//        words.add(new Word("green", "chokokki", R.drawable.color_green, R.raw.color_green));
+//        words.add(new Word("brown", "takaakki", R.drawable.color_brown, R.raw.color_brown));
+//        words.add(new Word("gray", "topoppi", R.drawable.color_gray, R.raw.color_gray));
+//        words.add(new Word("black", "kululli", R.drawable.color_black, R.raw.color_black));
+//        words.add(new Word("white", "kelelli", R.drawable.color_white, R.raw.color_white));
+//        words.add(new Word("dusty yellow", "topiise", R.drawable.color_dusty_yellow, R.raw.color_dusty_yellow));
+//        words.add(new Word("mustard yellow", "chiwiite", R.drawable.color_mustard_yellow, R.raw.color_mustard_yellow));
+//
+//        WordAdapter adapter = new WordAdapter(this, words, R.color.category_colors);
+//
+//        ListView listView = (ListView) findViewById(R.id.list);
+//
+//        listView.setAdapter(adapter);
+//
+//        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+//            @Override
+//            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
+//                Word word = words.get(position);
+//
+//                releaseMediaPlayer();
+//
+//                mMediaPlayer = MediaPlayer.create(ColorsActivity.this, word.getmAudioResourceId());
+//                mMediaPlayer.start(); // no need to call prepare(); create() does that for you
+//
+//                mMediaPlayer.setOnCompletionListener(mCompletionListener);
+//            }
+//        });
+//    }
+//
+//    @Override
+//    protected void onStop() {
+//        super.onStop();
+//        // When the activity is stopped, release the media player resources because we won't
+//        // be playing any more sounds.
+//        releaseMediaPlayer();
+//    }
+//
+//    /**
+//     * Clean up the media player by releasing its resources.
+//     */
+//    private void releaseMediaPlayer() {
+//        // If the media player is not null, then it may be currently playing a sound.
+//        if (mMediaPlayer != null) {
+//            // Regardless of the current state of the media player, release its resources
+//            // because we no longer need it.
+//            mMediaPlayer.release();
+//
+//            // Set the media player back to null. For our code, we've decided that
+//            // setting the media player to null is an easy way to tell that the media player
+//            // is not configured to play an audio file at the moment.
+//            mMediaPlayer = null;
+//        }
     }
 }
