+++ app/src/main/java/com/example/android/miwok/FamilyActivity.java
@@ -27,79 +27,79 @@ import java.util.ArrayList;
 
 public class FamilyActivity extends AppCompatActivity {
 
-    private MediaPlayer mMediaPlayer;
-
-    private MediaPlayer.OnCompletionListener mCompletionListener =  new MediaPlayer.OnCompletionListener(){
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            releaseMediaPlayer();
-        }
-    };
+//    private MediaPlayer mMediaPlayer;
+//
+//    private MediaPlayer.OnCompletionListener mCompletionListener =  new MediaPlayer.OnCompletionListener(){
+//        @Override
+//        public void onCompletion(MediaPlayer mediaPlayer) {
+//            releaseMediaPlayer();
+//        }
+//    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.word_list);
 
-        //Create an array of words
-        final ArrayList<Word> words = new ArrayList<Word>();
-
-//        words.add("one");
-        words.add(new Word("father", "әpә", R.drawable.family_father, R.raw.family_father));
-        words.add(new Word("mother", "әṭa", R.drawable.family_mother, R.raw.family_mother));
-        words.add(new Word("son", "angsi", R.drawable.family_son, R.raw.family_son));
-        words.add(new Word("daughter", "tune", R.drawable.family_daughter, R.raw.family_daughter));
-        words.add(new Word("older brother", "taachi", R.drawable.family_older_brother, R.raw.family_older_brother));
-        words.add(new Word("younger brother", "chalitti", R.drawable.family_younger_brother, R.raw.family_younger_brother));
-        words.add(new Word("older sister", "teṭe", R.drawable.family_older_sister, R.raw.family_older_sister));
-        words.add(new Word("younger sister", "kolliti", R.drawable.family_younger_sister, R.raw.family_younger_sister));
-        words.add(new Word( "grandmother", "ama", R.drawable.family_grandmother, R.raw.family_grandmother));
-        words.add(new Word("grandfather", "paapa", R.drawable.family_grandfather, R.raw.family_grandfather));
-
-        WordAdapter adapter = new WordAdapter(this, words, R.color.category_family);
-
-        ListView listView = (ListView) findViewById(R.id.list);
-
-        listView.setAdapter(adapter);
-
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
-                Word word = words.get(position);
-
-                releaseMediaPlayer();
-
-                mMediaPlayer = MediaPlayer.create(FamilyActivity.this, word.getmAudioResourceId());
-                mMediaPlayer.start(); // no need to call prepare(); create() does that for you
-
-                mMediaPlayer.setOnCompletionListener(mCompletionListener);
-            }
-        });
-
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        // When the activity is stopped, release the media player resources because we won't
-        // be playing any more sounds.
-        releaseMediaPlayer();
-    }
-
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-        }
+//        //Create an array of words
+//        final ArrayList<Word> words = new ArrayList<Word>();
+//
+////        words.add("one");
+//        words.add(new Word("father", "әpә", R.drawable.family_father, R.raw.family_father));
+//        words.add(new Word("mother", "әṭa", R.drawable.family_mother, R.raw.family_mother));
+//        words.add(new Word("son", "angsi", R.drawable.family_son, R.raw.family_son));
+//        words.add(new Word("daughter", "tune", R.drawable.family_daughter, R.raw.family_daughter));
+//        words.add(new Word("older brother", "taachi", R.drawable.family_older_brother, R.raw.family_older_brother));
+//        words.add(new Word("younger brother", "chalitti", R.drawable.family_younger_brother, R.raw.family_younger_brother));
+//        words.add(new Word("older sister", "teṭe", R.drawable.family_older_sister, R.raw.family_older_sister));
+//        words.add(new Word("younger sister", "kolliti", R.drawable.family_younger_sister, R.raw.family_younger_sister));
+//        words.add(new Word( "grandmother", "ama", R.drawable.family_grandmother, R.raw.family_grandmother));
+//        words.add(new Word("grandfather", "paapa", R.drawable.family_grandfather, R.raw.family_grandfather));
+//
+//        WordAdapter adapter = new WordAdapter(this, words, R.color.category_family);
+//
+//        ListView listView = (ListView) findViewById(R.id.list);
+//
+//        listView.setAdapter(adapter);
+//
+//        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+//            @Override
+//            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
+//                Word word = words.get(position);
+//
+//                releaseMediaPlayer();
+//
+//                mMediaPlayer = MediaPlayer.create(FamilyActivity.this, word.getmAudioResourceId());
+//                mMediaPlayer.start(); // no need to call prepare(); create() does that for you
+//
+//                mMediaPlayer.setOnCompletionListener(mCompletionListener);
+//            }
+//        });
+//
+//    }
+//
+//    @Override
+//    protected void onStop() {
+//        super.onStop();
+//        // When the activity is stopped, release the media player resources because we won't
+//        // be playing any more sounds.
+//        releaseMediaPlayer();
+//    }
+//
+//    /**
+//     * Clean up the media player by releasing its resources.
+//     */
+//    private void releaseMediaPlayer() {
+//        // If the media player is not null, then it may be currently playing a sound.
+//        if (mMediaPlayer != null) {
+//            // Regardless of the current state of the media player, release its resources
+//            // because we no longer need it.
+//            mMediaPlayer.release();
+//
+//            // Set the media player back to null. For our code, we've decided that
+//            // setting the media player to null is an easy way to tell that the media player
+//            // is not configured to play an audio file at the moment.
+//            mMediaPlayer = null;
+//        }
     }
 }
