+++ app/src/main/java/com/example/android/miwok/NumbersActivity.java
@@ -30,158 +30,124 @@ import java.util.ArrayList;
 
 public class NumbersActivity extends AppCompatActivity {
 
-    private MediaPlayer mMediaPlayer;
 
-    /** Handles audio focus when playing a sound file */
-    private AudioManager mAudioManager;
-
-    private MediaPlayer.OnCompletionListener mCompletionListener =  new MediaPlayer.OnCompletionListener(){
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            releaseMediaPlayer();
-        }
-    };
-
-    /**
-     * This listener gets triggered whenever the audio focus changes
-     * (i.e., we gain or lose audio focus because of another app or device).
-     */
-    private AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
-        @Override
-        public void onAudioFocusChange(int focusChange) {
-            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT ||
-                    focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK) {
-                // The AUDIOFOCUS_LOSS_TRANSIENT case means that we've lost audio focus for a
-                // short amount of time. The AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK case means that
-                // our app is allowed to continue playing sound but at a lower volume. We'll treat
-                // both cases the same way because our app is playing short sound files.
-
-                // Pause playback and reset player to the start of the file. That way, we can
-                // play the word from the beginning when we resume playback.
-                mMediaPlayer.pause();
-                mMediaPlayer.seekTo(0);
-            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                // The AUDIOFOCUS_GAIN case means we have regained focus and can resume playback.
-                mMediaPlayer.start();
-            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                // The AUDIOFOCUS_LOSS case means we've lost audio focus and
-                // Stop playback and clean up resources
-                releaseMediaPlayer();
-            }
-        }
-    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.word_list);
 
-        // Create and setup the {@link AudioManager} to request audio focus
-        mAudioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.container, new NumbersFragment())
+                .commit();
 
-        //Create an array of words
-        final ArrayList<Word> words = new ArrayList<Word>();
-
-//        words.add("one");
-        words.add(new Word("one", "lutti", R.drawable.number_one, R.raw.number_one));
-        words.add(new Word("two", "otiiko", R.drawable.number_two, R.raw.number_two));
-        words.add(new Word("three", "tolookosu", R.drawable.number_three, R.raw.number_three));
-        words.add(new Word("four", "oyyisa", R.drawable.number_four, R.raw.number_four));
-        words.add(new Word("five", "massokka", R.drawable.number_five, R.raw.number_five));
-        words.add(new Word("six", "temmokka", R.drawable.number_six, R.raw.number_six));
-        words.add(new Word("seven", "kenekaku", R.drawable.number_seven, R.raw.number_seven));
-        words.add(new Word("eight", "kawinta", R.drawable.number_eight, R.raw.number_eight));
-        words.add(new Word( "nine", "wo'e", R.drawable.number_nine, R.raw.number_nine));
-        words.add(new Word("ten", "na'aacha", R.drawable.number_ten, R.raw.number_ten));
-
-
-        // WHILE LOOP
-
-//        int index = 0;
+//        // Create and setup the {@link AudioManager} to request audio focus
+//        mAudioManager = (AudioManager) getSystemService(Context.AUDIO_SERVICE);
 //
-//        while (index < words.size()) {
-//            LinearLayout rootView = (LinearLayout)findViewById(R.id.rootView);
-//            TextView wordView = new TextView(this);
-//            wordView.setText(words.get(index));
-//            rootView.addView(wordView);
+//        //Create an array of words
+//        final ArrayList<Word> words = new ArrayList<Word>();
 //
-//            //Update the counter variable
-//            index++;
+////        words.add("one");
+//        words.add(new Word("one", "lutti", R.drawable.number_one, R.raw.number_one));
+//        words.add(new Word("two", "otiiko", R.drawable.number_two, R.raw.number_two));
+//        words.add(new Word("three", "tolookosu", R.drawable.number_three, R.raw.number_three));
+//        words.add(new Word("four", "oyyisa", R.drawable.number_four, R.raw.number_four));
+//        words.add(new Word("five", "massokka", R.drawable.number_five, R.raw.number_five));
+//        words.add(new Word("six", "temmokka", R.drawable.number_six, R.raw.number_six));
+//        words.add(new Word("seven", "kenekaku", R.drawable.number_seven, R.raw.number_seven));
+//        words.add(new Word("eight", "kawinta", R.drawable.number_eight, R.raw.number_eight));
+//        words.add(new Word( "nine", "wo'e", R.drawable.number_nine, R.raw.number_nine));
+//        words.add(new Word("ten", "na'aacha", R.drawable.number_ten, R.raw.number_ten));
+//
+//
+//        // WHILE LOOP
+//
+////        int index = 0;
+////
+////        while (index < words.size()) {
+////            LinearLayout rootView = (LinearLayout)findViewById(R.id.rootView);
+////            TextView wordView = new TextView(this);
+////            wordView.setText(words.get(index));
+////            rootView.addView(wordView);
+////
+////            //Update the counter variable
+////            index++;
+////        }
+//
+//        // FOR LOOP
+//
+////        // Find the root view of the layout
+////        LinearLayout rootView = (LinearLayout)findViewById(R.id.rootView);
+////
+////        for (int index = 0; index < words.size(); index++) {
+////            // Create a new TextView
+////            TextView wordView = new TextView(this);
+////            //Set the text to be word at the current index
+////            wordView.setText(words.get(index));
+////            // Add this TextView as another child to the root view of this layout
+////            rootView.addView(wordView);
+////        }
+//
+//        WordAdapter adapter = new WordAdapter(this, words, R.color.category_numbers);
+//
+//        ListView listView = (ListView) findViewById(R.id.list);
+//
+//        listView.setAdapter(adapter);
+//
+//        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+//            @Override
+//            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
+//                Word word = words.get(position);
+//
+//                releaseMediaPlayer();
+//
+//                //Request audio focus for playback
+//                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
+//                        // Use the music stream
+//                        AudioManager.STREAM_MUSIC,
+//                        // Request permanent focus
+//                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
+//                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
+//
+//                    // We have audio focus now
+//
+//
+//                    mMediaPlayer = MediaPlayer.create(NumbersActivity.this, word.getmAudioResourceId());
+//                    mMediaPlayer.start(); // no need to call prepare(); create() does that for you
+//
+//                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
+//
+//                }
+//            }
+//        });
+//    }
+//
+//    @Override
+//    protected void onStop() {
+//        super.onStop();
+//        // When the activity is stopped, release the media player resources because we won't
+//        // be playing any more sounds.
+//        releaseMediaPlayer();
+//    }
+//
+//    /**
+//     * Clean up the media player by releasing its resources.
+//     */
+//    private void releaseMediaPlayer() {
+//        // If the media player is not null, then it may be currently playing a sound.
+//        if (mMediaPlayer != null) {
+//            // Regardless of the current state of the media player, release its resources
+//            // because we no longer need it.
+//            mMediaPlayer.release();
+//
+//            // Set the media player back to null. For our code, we've decided that
+//            // setting the media player to null is an easy way to tell that the media player
+//            // is not configured to play an audio file at the moment.
+//            mMediaPlayer = null;
+//
+//            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
 //        }
-
-        // FOR LOOP
-
-//        // Find the root view of the layout
-//        LinearLayout rootView = (LinearLayout)findViewById(R.id.rootView);
-//
-//        for (int index = 0; index < words.size(); index++) {
-//            // Create a new TextView
-//            TextView wordView = new TextView(this);
-//            //Set the text to be word at the current index
-//            wordView.setText(words.get(index));
-//            // Add this TextView as another child to the root view of this layout
-//            rootView.addView(wordView);
-//        }
-
-        WordAdapter adapter = new WordAdapter(this, words, R.color.category_numbers);
-
-        ListView listView = (ListView) findViewById(R.id.list);
-
-        listView.setAdapter(adapter);
-
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
-                Word word = words.get(position);
-
-                releaseMediaPlayer();
-
-                //Request audio focus for playback
-                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
-                        // Use the music stream
-                        AudioManager.STREAM_MUSIC,
-                        // Request permanent focus
-                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
-                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-
-                    // We have audio focus now
-
-
-                    mMediaPlayer = MediaPlayer.create(NumbersActivity.this, word.getmAudioResourceId());
-                    mMediaPlayer.start(); // no need to call prepare(); create() does that for you
-
-                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
-
-                }
-            }
-        });
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        // When the activity is stopped, release the media player resources because we won't
-        // be playing any more sounds.
-        releaseMediaPlayer();
-    }
-
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-
-            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-        }
     }
 
 }
